Renamed support to common and handled the imports and md file references
* Renamed support to common
* Modified imports in various files
* Modified references in .md files

Fixes : #213
