+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -15,7 +15,7 @@ Contributing Conventions:
 
 1. Include descriptive PR titles with [<component-name>] prepended.
 2. Build and test your changes before submitting a PR. 
-3. If logging then check the logging.md file in ops/support_ops
+3. If logging then check the logging.md file in common/
 4. Remember to check the linting issues beforehand as well to prevent your checks from failing.
 5. Remember to sign-off your commits
