+++ common/logging.md
