+++ common/ops/abstract_ops.py
